Use separate test database from local
This moves DATABASE_URL to config so it is easier to run tests in a separate
test database. This helps to ensure more consistent tests since the database is
destroyed and created as part of the test run.

This also adds a default for VCAP_APPLICATION, so no special variables are
necessary for local development.